requirements.txt
streamlit
pandas
numpy
matplotlib
supabase
python-dotenv